move anrwatchdog(bundled jar) to playstore flavor
