+++ app/src/foss/java/ohi/andre/consolelauncher/FlavorUtils.java
@@ -1,0 +1,9 @@
+package ohi.andre.consolelauncher;
+
+public class FlavorUtils {
+
+    public static void startANR () {
+
+    }
+	
+}
